requirements.txt
aniso8601==3.0.0
backports.ssl-match-hostname==3.5.0.1
bintrees==2.0.7
cachetools==2.0.1
ccxt==1.11.104
certifi==2018.1.18
cexio==0.1.5
chardet==3.0.4
click==6.7
configparser==3.5.0
enum34==1.1.6
firebase-admin==2.9.0
Flask==0.12.2
Flask-RESTful==0.3.6
future==0.16.0
futures==3.2.0
gdax==1.0.6
google-api-core==1.1.0
google-api-python-client==1.6.5
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-cloud-core==0.24.1
google-cloud-firestore==0.29.0
google-cloud-storage==1.1.1
google-resumable-media==0.0.2
googleapis-common-protos==1.5.3
httplib2==0.10.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
oauth2client==4.1.2
protobuf==3.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-firebase==1.2
pytz==2018.3
requests==2.18.4
rsa==3.4.2
simplejson==3.13.2
six==1.11.0
stripe==1.79.1
uritemplate==3.0.0
urllib3==1.22
webapp2==2.5.2
websocket-client==0.40.0
Werkzeug==0.14.1
